introduce REST Endpoint abstractions
