+++ src/Tickets/Commerce/Gateways/PayPal/REST/On_Boarding_Endpoint.php
@@ -2,6 +2,7 @@
 
 namespace TEC\Tickets\Commerce\Gateways\PayPal\REST;
 
+use TEC\Tickets\Commerce\Gateways\Abstract_REST_Endpoint;
 use TEC\Tickets\Commerce\Gateways\PayPal\Client;
 use TEC\Tickets\Commerce\Gateways\PayPal\Merchant;
 use TEC\Tickets\Commerce\Gateways\PayPal\Refresh_Token;
@@ -10,7 +11,6 @@ use TEC\Tickets\Commerce\Gateways\PayPal\Signup;
 use TEC\Tickets\Commerce\Gateways\PayPal\Webhooks;
 use TEC\Tickets\Commerce\Gateways\PayPal\WhoDat;
 use TEC\Tickets\Commerce\Notice_Handler;
-use Tribe__Documentation__Swagger__Provider_Interface;
 use Tribe__Settings;
 use Tribe__Utils__Array as Arr;
 
@@ -27,21 +27,15 @@ use WP_REST_Server;
  *
  * @package TEC\Tickets\Commerce\Gateways\PayPal\REST
  */
-class On_Boarding_Endpoint implements Tribe__Documentation__Swagger__Provider_Interface {
+class On_Boarding_Endpoint extends Abstract_REST_Endpoint {
 
 	/**
-	 * The REST API endpoint path.
-	 *
-	 * @since 5.1.9
-	 *
-	 * @var string
+	 * @inheritDoc
 	 */
 	protected $path = '/commerce/paypal/on-boarding';
 
 	/**
-	 * Register the actual endpoint on WP Rest API.
-	 *
-	 * @since 5.1.9
+	 * @inheritDoc
 	 */
 	public function register() {
 		$namespace     = tribe( 'tickets.rest-v1.main' )->get_events_route_namespace();
@@ -70,45 +64,6 @@ class On_Boarding_Endpoint implements Tribe__Documentation__Swagger__Provider_In
 		);
 
 		$documentation->register_documentation_provider( $this->get_endpoint_path(), $this );
-	}
-
-	/**
-	 * Gets the Endpoint path for the on boarding process.
-	 *
-	 * @since 5.1.9
-	 *
-	 * @return string
-	 */
-	public function get_endpoint_path() {
-		return $this->path;
-	}
-
-	/**
-	 * Get the REST API route URL.
-	 *
-	 * @since 5.1.9
-	 *
-	 * @return string The REST API route URL.
-	 */
-	public function get_route_url() {
-		$namespace = tribe( 'tickets.rest-v1.main' )->get_events_route_namespace();
-
-		return rest_url( '/' . $namespace . $this->get_endpoint_path(), 'https' );
-	}
-
-	/**
-	 * Gets the Return URL pointing to this on boarding route.
-	 *
-	 * @since 5.1.9
-	 *
-	 * @return string
-	 */
-	public function get_return_url( $hash = null ) {
-		$arguments = [
-			'hash' => $hash,
-		];
-
-		return add_query_arg( $arguments, $this->get_route_url() );
 	}
 
 	/**
@@ -414,35 +369,5 @@ class On_Boarding_Endpoint implements Tribe__Documentation__Swagger__Provider_In
 				'sanitize_callback' => [ $this, 'sanitize_callback' ],
 			],
 		];
-	}
-
-	/**
-	 * Sanitize a request argument based on details registered to the route.
-	 *
-	 * @since 5.1.9
-	 *
-	 * @param mixed $value Value of the 'filter' argument.
-	 *
-	 * @return string|array
-	 */
-	public function sanitize_callback( $value ) {
-		if ( is_array( $value ) ) {
-			return array_map( 'sanitize_text_field', $value );
-		}
-
-		return sanitize_text_field( $value );
-	}
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * @TODO  We need to make sure Swagger documentation is present.
-	 *
-	 * @since 5.1.9
-	 *
-	 * @return array
-	 */
-	public function get_documentation() {
-		return [];
 	}
 }
