+++ src/Tickets/Commerce/Gateways/PayPal/REST/Order_Endpoint.php
@@ -2,29 +2,17 @@
 
 namespace TEC\Tickets\Commerce\Gateways\PayPal\REST;
 
-use tad\WPBrowser\Adapters\WP;
 use TEC\Tickets\Commerce\Cart;
+use TEC\Tickets\Commerce\Gateways\Abstract_REST_Endpoint;
 use TEC\Tickets\Commerce\Gateways\PayPal\Gateway;
 use TEC\Tickets\Commerce\Gateways\PayPal\Status;
 use TEC\Tickets\Commerce\Order;
 
 use TEC\Tickets\Commerce\Gateways\PayPal\Client;
-use TEC\Tickets\Commerce\Gateways\PayPal\Merchant;
-use TEC\Tickets\Commerce\Gateways\PayPal\Refresh_Token;
-
-use TEC\Tickets\Commerce\Gateways\PayPal\Signup;
-use TEC\Tickets\Commerce\Gateways\PayPal\WhoDat;
-
-
-use TEC\Tickets\Commerce\Module;
 use TEC\Tickets\Commerce\Status\Denied;
 use TEC\Tickets\Commerce\Status\Pending;
-use TEC\Tickets\Commerce\Status\Completed;
-use TEC\Tickets\Commerce\Status\Created;
 use TEC\Tickets\Commerce\Status\Status_Handler;
 use TEC\Tickets\Commerce\Success;
-use Tribe__Documentation__Swagger__Provider_Interface;
-use Tribe__Settings;
 use Tribe__Utils__Array as Arr;
 
 use WP_Error;
@@ -40,7 +28,7 @@ use WP_REST_Server;
  *
  * @package TEC\Tickets\Commerce\Gateways\PayPal\REST
  */
-class Order_Endpoint implements Tribe__Documentation__Swagger__Provider_Interface {
+class Order_Endpoint extends Abstract_REST_Endpoint {
 
 	/**
 	 * The REST API endpoint path.
@@ -94,30 +82,6 @@ class Order_Endpoint implements Tribe__Documentation__Swagger__Provider_Interfac
 		);
 
 		$documentation->register_documentation_provider( $this->get_endpoint_path(), $this );
-	}
-
-	/**
-	 * Gets the Endpoint path for the on boarding process.
-	 *
-	 * @since 5.1.9
-	 *
-	 * @return string
-	 */
-	public function get_endpoint_path() {
-		return $this->path;
-	}
-
-	/**
-	 * Get the REST API route URL.
-	 *
-	 * @since 5.1.9
-	 *
-	 * @return string The REST API route URL.
-	 */
-	public function get_route_url() {
-		$namespace = tribe( 'tickets.rest-v1.main' )->get_events_route_namespace();
-
-		return rest_url( '/' . $namespace . $this->get_endpoint_path(), 'https' );
 	}
 
 	/**
@@ -432,19 +396,6 @@ class Order_Endpoint implements Tribe__Documentation__Swagger__Provider_Interfac
 		}
 
 		return sanitize_text_field( $value );
-	}
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * @TODO  We need to make sure Swagger documentation is present.
-	 *
-	 * @since 5.1.9
-	 *
-	 * @return array
-	 */
-	public function get_documentation() {
-		return [];
 	}
 
 	/**
